Sort apps/components in the bottomsheet
