+++ app-compose/build.gradle.kts
@@ -96,7 +96,6 @@ dependencies {
     implementation(projects.feature.search)
     implementation(projects.feature.settings)
     implementation(projects.feature.ruledetail)
-    implementation(projects.feature.sort)
 
     implementation(projects.core.analytics)
     implementation(projects.core.ui)
