+++ app-compose/src/main/java/com/merxury/blocker/navigation/BlockerNavHost.kt
@@ -36,10 +36,6 @@ import com.merxury.blocker.feature.ruledetail.navigation.ruleDetailScreen
 import com.merxury.blocker.feature.search.navigation.searchScreen
 import com.merxury.blocker.feature.settings.navigation.navigateToSettings
 import com.merxury.blocker.feature.settings.navigation.settingsScreen
-import com.merxury.blocker.feature.sort.navigation.appSortScreen
-import com.merxury.blocker.feature.sort.navigation.componentSortScreen
-import com.merxury.blocker.feature.sort.navigation.navigateToAppSortScreen
-import com.merxury.blocker.feature.sort.navigation.navigateToComponentSortScreen
 
 /**
  * Top-level navigation graph. Navigation is organized as explained at
@@ -66,13 +62,11 @@ fun BlockerNavHost(
             navigateToAppDetail = navController::navigateToAppDetail,
             navigateToSettings = navController::navigateToSettings,
             navigateToSupportAndFeedback = navController::navigateToSupportAndFeedback,
-            navigateTooAppSortScreen = navController::navigateToAppSortScreen,
         )
         detailScreen(
             onBackClick = onBackClick,
             snackbarHostState = snackbarHostState,
             navigateToComponentDetail = navController::navigateToComponentDetail,
-            navigatedToComponentSortScreen = navController::navigateToComponentSortScreen,
         )
         generalRuleScreen(
             navigateToRuleDetail = navController::navigateToRuleDetail,
@@ -91,12 +85,6 @@ fun BlockerNavHost(
         )
         supportAndFeedbackScreen(onBackClick)
         componentDetailScreen(
-            dismissHandler = onBackClick,
-        )
-        componentSortScreen(
-            dismissHandler = onBackClick,
-        )
-        appSortScreen(
             dismissHandler = onBackClick,
         )
     }
