+++ core/ui/src/main/java/com/merxury/blocker/core/ui/bottomsheet/AppSortBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.merxury.blocker.feature.sort/*
+/*
  * Copyright 2023 Blocker
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,29 +13,21 @@ package com.merxury.blocker.feature.sort/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.merxury.blocker.core.ui.bottomsheet
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.merxury.blocker.core.designsystem.component.ItemHeader
 import com.merxury.blocker.core.designsystem.segmentedbuttons.SegmentedButtons
 import com.merxury.blocker.core.designsystem.theme.BlockerTheme
@@ -44,42 +36,12 @@ import com.merxury.blocker.core.model.preference.AppSorting.NAME
 import com.merxury.blocker.core.model.preference.SortingOrder
 import com.merxury.blocker.core.model.preference.SortingOrder.ASCENDING
 import com.merxury.blocker.core.model.preference.SortingOrder.DESCENDING
+import com.merxury.blocker.core.ui.R
+import com.merxury.blocker.core.ui.bottomsheet.AppSortInfoUiState.Loading
+import com.merxury.blocker.core.ui.bottomsheet.AppSortInfoUiState.Success
 import com.merxury.blocker.core.ui.screen.LoadingScreen
-import com.merxury.blocker.feature.sort.viewmodel.AppSortInfo
-import com.merxury.blocker.feature.sort.viewmodel.AppSortInfoUiState
-import com.merxury.blocker.feature.sort.viewmodel.AppSortInfoUiState.Loading
-import com.merxury.blocker.feature.sort.viewmodel.AppSortInfoUiState.Success
-import com.merxury.blocker.feature.sort.viewmodel.AppSortViewModel
 import java.lang.Boolean.FALSE
 import java.lang.Boolean.TRUE
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AppSortBottomSheetRoute(
-    dismissHandler: () -> Unit,
-    modifier: Modifier = Modifier,
-    viewModel: AppSortViewModel = hiltViewModel(),
-) {
-    val uiState by viewModel.appSortInfoUiState.collectAsStateWithLifecycle()
-    val skipPartiallyExpanded by remember { mutableStateOf(false) }
-    val bottomSheetState = rememberModalBottomSheetState(
-        skipPartiallyExpanded = skipPartiallyExpanded,
-    )
-    ModalBottomSheet(
-        onDismissRequest = { dismissHandler() },
-        sheetState = bottomSheetState,
-        containerColor = MaterialTheme.colorScheme.surfaceVariant,
-        scrimColor = Color.Transparent,
-    ) {
-        AppSortBottomSheet(
-            uiState = uiState,
-            modifier = modifier,
-            onSortByClick = viewModel::updateAppSorting,
-            onSortOrderClick = viewModel::updateAppSortingOrder,
-            onChangeShowRunningAppsOnTop = viewModel::updateShowRunningAppsOnTop,
-        )
-    }
-}
 
 @Composable
 fun AppSortBottomSheet(
