+++ core/ui/src/main/java/com/merxury/blocker/core/ui/bottomsheet/ComponentSortBottomSheet.kt
@@ -14,30 +14,21 @@
  * limitations under the License.
  */
 
-package com.merxury.blocker.feature.sort
+package com.merxury.blocker.core.ui.bottomsheet
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.merxury.blocker.core.designsystem.component.ItemHeader
 import com.merxury.blocker.core.designsystem.segmentedbuttons.SegmentedButtons
 import com.merxury.blocker.core.designsystem.theme.BlockerTheme
@@ -51,40 +42,10 @@ import com.merxury.blocker.core.model.preference.ComponentSorting.PACKAGE_NAME
 import com.merxury.blocker.core.model.preference.SortingOrder
 import com.merxury.blocker.core.model.preference.SortingOrder.ASCENDING
 import com.merxury.blocker.core.model.preference.SortingOrder.DESCENDING
+import com.merxury.blocker.core.ui.R
+import com.merxury.blocker.core.ui.bottomsheet.ComponentSortInfoUiState.Loading
+import com.merxury.blocker.core.ui.bottomsheet.ComponentSortInfoUiState.Success
 import com.merxury.blocker.core.ui.screen.LoadingScreen
-import com.merxury.blocker.feature.sort.viewmodel.ComponentSortInfo
-import com.merxury.blocker.feature.sort.viewmodel.ComponentSortInfoUiState
-import com.merxury.blocker.feature.sort.viewmodel.ComponentSortInfoUiState.Loading
-import com.merxury.blocker.feature.sort.viewmodel.ComponentSortInfoUiState.Success
-import com.merxury.blocker.feature.sort.viewmodel.SortViewModel
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ComponentSortBottomSheetRoute(
-    dismissHandler: () -> Unit,
-    modifier: Modifier = Modifier,
-    viewModel: SortViewModel = hiltViewModel(),
-) {
-    val uiState by viewModel.componentSortInfoUiState.collectAsStateWithLifecycle()
-    val skipPartiallyExpanded by remember { mutableStateOf(false) }
-    val bottomSheetState = rememberModalBottomSheetState(
-        skipPartiallyExpanded = skipPartiallyExpanded,
-    )
-    ModalBottomSheet(
-        onDismissRequest = { dismissHandler() },
-        sheetState = bottomSheetState,
-        containerColor = MaterialTheme.colorScheme.surfaceVariant,
-        scrimColor = Color.Transparent,
-    ) {
-        ComponentSortBottomSheet(
-            uiState = uiState,
-            modifier = modifier,
-            onSortByClick = viewModel::updateComponentSorting,
-            onSortOrderClick = viewModel::updateComponentSortingOrder,
-            onShowPriorityClick = viewModel::updateComponentShowPriority,
-        )
-    }
-}
 
 @Composable
 fun ComponentSortBottomSheet(
