+++ feature/appdetail/src/main/java/com/merxury/blocker/feature/appdetail/AppDetailScreen.kt
@@ -61,6 +61,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Velocity
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.dokar.sheets.BottomSheet
+import com.dokar.sheets.rememberBottomSheetState
 import com.merxury.blocker.core.designsystem.component.BlockerCollapsingTopAppBar
 import com.merxury.blocker.core.designsystem.component.BlockerErrorAlertDialog
 import com.merxury.blocker.core.designsystem.component.BlockerScrollableTabRow
@@ -69,6 +71,9 @@ import com.merxury.blocker.core.designsystem.component.BlockerTab
 import com.merxury.blocker.core.designsystem.component.MaxToolbarHeight
 import com.merxury.blocker.core.designsystem.component.MinToolbarHeight
 import com.merxury.blocker.core.designsystem.theme.BlockerTheme
+import com.merxury.blocker.core.model.preference.ComponentShowPriority
+import com.merxury.blocker.core.model.preference.ComponentSorting
+import com.merxury.blocker.core.model.preference.SortingOrder
 import com.merxury.blocker.core.rule.entity.RuleWorkResult.CANCELLED
 import com.merxury.blocker.core.rule.entity.RuleWorkResult.FINISHED
 import com.merxury.blocker.core.rule.entity.RuleWorkResult.FOLDER_NOT_DEFINED
@@ -85,6 +90,9 @@ import com.merxury.blocker.core.ui.AppDetailTabs.Service
 import com.merxury.blocker.core.ui.TabState
 import com.merxury.blocker.core.ui.TrackScreenViewEvent
 import com.merxury.blocker.core.ui.applist.model.AppItem
+import com.merxury.blocker.core.ui.bottomsheet.ComponentSortBottomSheet
+import com.merxury.blocker.core.ui.bottomsheet.ComponentSortInfo
+import com.merxury.blocker.core.ui.bottomsheet.ComponentSortInfoUiState
 import com.merxury.blocker.core.ui.component.ComponentList
 import com.merxury.blocker.core.ui.screen.ErrorScreen
 import com.merxury.blocker.core.ui.screen.LoadingScreen
@@ -107,7 +115,6 @@ import com.merxury.blocker.core.rule.R.string as rulestring
 fun AppDetailRoute(
     onBackClick: () -> Unit,
     navigateToComponentDetail: (String) -> Unit,
-    navigatedToComponentSortScreen: () -> Unit,
     snackbarHostState: SnackbarHostState,
     modifier: Modifier = Modifier,
     viewModel: AppDetailViewModel = hiltViewModel(),
@@ -117,6 +124,7 @@ fun AppDetailRoute(
     val errorState by viewModel.errorState.collectAsStateWithLifecycle()
     val topAppBarUiState by viewModel.appBarUiState.collectAsStateWithLifecycle()
     val componentListUiState by viewModel.componentListUiState.collectAsStateWithLifecycle()
+    val bottomSheetState by viewModel.componentSortInfoUiState.collectAsStateWithLifecycle()
     val event by viewModel.eventFlow.collectAsState(initial = null)
     val scope = rememberCoroutineScope()
     val clipboardManager = LocalClipboardManager.current
@@ -126,6 +134,7 @@ fun AppDetailRoute(
         topAppBarUiState = topAppBarUiState,
         componentListUiState = componentListUiState,
         tabState = tabState,
+        bottomSheetState = bottomSheetState,
         navigateToComponentDetail = navigateToComponentDetail,
         modifier = modifier.fillMaxSize(),
         onLaunchAppClick = { packageName ->
@@ -147,7 +156,9 @@ fun AppDetailRoute(
         onLaunchActivityClick = viewModel::launchActivity,
         onCopyNameClick = { clipboardManager.setText(AnnotatedString(it)) },
         onCopyFullNameClick = { clipboardManager.setText(AnnotatedString(it)) },
-        navigatedToComponentSortScreen = navigatedToComponentSortScreen,
+        onSortByClick = viewModel::updateComponentSorting,
+        onSortOrderClick = viewModel::updateComponentSortingOrder,
+        onShowPriorityClick = viewModel::updateComponentShowPriority,
     )
     if (errorState != null) {
         BlockerErrorAlertDialog(
@@ -192,6 +203,7 @@ fun AppDetailScreen(
     topAppBarUiState: AppBarUiState,
     componentListUiState: ComponentListUiState,
     tabState: TabState<AppDetailTabs>,
+    bottomSheetState: ComponentSortInfoUiState,
     onBackClick: () -> Unit,
     onLaunchAppClick: (String) -> Unit,
     switchTab: (AppDetailTabs) -> Unit,
@@ -211,7 +223,9 @@ fun AppDetailScreen(
     onLaunchActivityClick: (String, String) -> Unit = { _, _ -> },
     onCopyNameClick: (String) -> Unit = { _ -> },
     onCopyFullNameClick: (String) -> Unit = { _ -> },
-    navigatedToComponentSortScreen: () -> Unit,
+    onSortByClick: (ComponentSorting) -> Unit = {},
+    onSortOrderClick: (SortingOrder) -> Unit = {},
+    onShowPriorityClick: (ComponentShowPriority) -> Unit = {},
 ) {
     when (appInfoUiState) {
         is AppInfoUiState.Loading -> {
@@ -224,6 +238,7 @@ fun AppDetailScreen(
                 topAppBarUiState = topAppBarUiState,
                 componentListUiState = componentListUiState,
                 tabState = tabState,
+                bottomSheetState = bottomSheetState,
                 onBackClick = onBackClick,
                 navigateToComponentDetail = navigateToComponentDetail,
                 onLaunchAppClick = onLaunchAppClick,
@@ -243,7 +258,9 @@ fun AppDetailScreen(
                 onLaunchActivityClick = onLaunchActivityClick,
                 onCopyNameClick = onCopyNameClick,
                 onCopyFullNameClick = onCopyFullNameClick,
-                navigatedToComponentSortScreen = navigatedToComponentSortScreen,
+                onSortByClick = onSortByClick,
+                onSortOrderClick = onSortOrderClick,
+                onShowPriorityClick = onShowPriorityClick,
             )
         }
 
@@ -257,6 +274,7 @@ fun AppDetailContent(
     app: AppItem,
     tabState: TabState<AppDetailTabs>,
     componentListUiState: ComponentListUiState,
+    bottomSheetState: ComponentSortInfoUiState,
     onBackClick: () -> Unit,
     onLaunchAppClick: (String) -> Unit,
     switchTab: (AppDetailTabs) -> Unit,
@@ -277,8 +295,11 @@ fun AppDetailContent(
     onLaunchActivityClick: (String, String) -> Unit = { _, _ -> },
     onCopyNameClick: (String) -> Unit = { _ -> },
     onCopyFullNameClick: (String) -> Unit = { _ -> },
-    navigatedToComponentSortScreen: () -> Unit,
+    onSortByClick: (ComponentSorting) -> Unit = {},
+    onSortOrderClick: (SortingOrder) -> Unit = {},
+    onShowPriorityClick: (ComponentShowPriority) -> Unit = {},
 ) {
+    val sheetState = rememberBottomSheetState()
     val listState = rememberLazyListState()
     val systemStatusHeight = WindowInsets.systemBars.asPaddingValues().calculateTopPadding()
     val toolbarHeightRange = with(LocalDensity.current) {
@@ -328,7 +349,11 @@ fun AppDetailContent(
                         onSearchModeChange = onSearchModeChanged,
                         blockAllComponents = blockAllComponents,
                         enableAllComponents = enableAllComponents,
-                        navigatedToComponentSortScreen = navigatedToComponentSortScreen,
+                        navigatedToComponentSortScreen = {
+                            scope.launch {
+                                sheetState.expand()
+                            }
+                        },
                     )
                 },
                 subtitle = app.packageName,
@@ -370,6 +395,14 @@ fun AppDetailContent(
             onLaunchActivityClick = onLaunchActivityClick,
             onCopyNameClick = onCopyNameClick,
             onCopyFullNameClick = onCopyFullNameClick,
+        )
+    }
+    BottomSheet(state = sheetState) {
+        ComponentSortBottomSheet(
+            uiState = bottomSheetState,
+            onSortByClick = onSortByClick,
+            onSortOrderClick = onSortOrderClick,
+            onShowPriorityClick = onShowPriorityClick,
         )
     }
 }
@@ -565,6 +598,9 @@ fun AppDetailScreenPreview() {
         Surface {
             AppDetailScreen(
                 appInfoUiState = Success(appInfo = app),
+                bottomSheetState = ComponentSortInfoUiState.Success(
+                    ComponentSortInfo(),
+                ),
                 componentListUiState = ComponentListUiState(),
                 tabState = tabState,
                 onLaunchAppClick = {},
@@ -573,7 +609,6 @@ fun AppDetailScreenPreview() {
                 topAppBarUiState = AppBarUiState(),
                 onSearchTextChanged = {},
                 onSearchModeChanged = {},
-                navigatedToComponentSortScreen = {},
             )
         }
     }
@@ -612,6 +647,9 @@ fun AppDetailScreenCollapsedPreview() {
         Surface {
             AppDetailScreen(
                 appInfoUiState = Success(appInfo = app),
+                bottomSheetState = ComponentSortInfoUiState.Success(
+                    ComponentSortInfo(),
+                ),
                 componentListUiState = ComponentListUiState(),
                 tabState = tabState,
                 onLaunchAppClick = {},
@@ -620,7 +658,6 @@ fun AppDetailScreenCollapsedPreview() {
                 topAppBarUiState = AppBarUiState(),
                 onSearchTextChanged = {},
                 onSearchModeChanged = {},
-                navigatedToComponentSortScreen = {},
             )
         }
     }
