+++ feature/appdetail/src/main/java/com/merxury/blocker/feature/appdetail/navigation/AppDetailNavigation.kt
@@ -69,7 +69,6 @@ fun NavGraphBuilder.detailScreen(
     onBackClick: () -> Unit,
     snackbarHostState: SnackbarHostState,
     navigateToComponentDetail: (String) -> Unit,
-    navigatedToComponentSortScreen: () -> Unit,
 ) {
     composable(
         route = "app_detail_route/{$packageNameArg}?screen={$tabArg}?keyword={$keywordArg}",
@@ -83,7 +82,6 @@ fun NavGraphBuilder.detailScreen(
             onBackClick = onBackClick,
             snackbarHostState = snackbarHostState,
             navigateToComponentDetail = navigateToComponentDetail,
-            navigatedToComponentSortScreen = navigatedToComponentSortScreen,
         )
     }
 }
