+++ feature/applist/src/main/java/com/merxury/blocker/feature/applist/AppListViewModel.kt
@@ -40,6 +40,9 @@ import com.merxury.blocker.core.model.preference.SortingOrder
 import com.merxury.blocker.core.result.Result
 import com.merxury.blocker.core.ui.applist.model.AppItem
 import com.merxury.blocker.core.ui.applist.model.toAppServiceStatus
+import com.merxury.blocker.core.ui.bottomsheet.AppSortInfo
+import com.merxury.blocker.core.ui.bottomsheet.AppSortInfoUiState
+import com.merxury.blocker.core.ui.bottomsheet.AppSortInfoUiState.Loading
 import com.merxury.blocker.core.ui.data.UiMessage
 import com.merxury.blocker.core.ui.data.toErrorMessage
 import com.merxury.blocker.core.ui.state.AppStateCache
@@ -76,6 +79,9 @@ class AppListViewModel @Inject constructor(
     @Dispatcher(DEFAULT) private val cpuDispatcher: CoroutineDispatcher,
     private val analyticsHelper: AnalyticsHelper,
 ) : AndroidViewModel(app) {
+    private val _appSortInfoUiState: MutableStateFlow<AppSortInfoUiState> =
+        MutableStateFlow(Loading)
+    val appSortInfoUiState = _appSortInfoUiState.asStateFlow()
     private val _uiState = MutableStateFlow<AppListUiState>(Initializing())
     val uiState = _uiState.asStateFlow()
     private val _errorState = MutableStateFlow<UiMessage?>(null)
@@ -94,6 +100,7 @@ class AppListViewModel @Inject constructor(
 
     init {
         loadData()
+        loadAppSortInfo()
         updateInstalledAppList()
         listenSortingChanges()
         listenShowRunningAppsOnTopChanges()
@@ -160,6 +167,30 @@ class AppListViewModel @Inject constructor(
     fun filter(keyword: String) {
         currentSearchKeyword = keyword
         loadData()
+    }
+
+    private fun loadAppSortInfo() = viewModelScope.launch {
+        val userData = userDataRepository.userData.first()
+        val sorting = userData.appSorting
+        val order = userData.appSortingOrder
+        val showRunningAppsOnTop = userData.showRunningAppsOnTop
+        _appSortInfoUiState.emit(
+            AppSortInfoUiState.Success(AppSortInfo(sorting, order, showRunningAppsOnTop)),
+        )
+    }
+
+    fun updateAppSorting(sorting: AppSorting) = viewModelScope.launch {
+        userDataRepository.setAppSorting(sorting)
+    }
+
+    fun updateAppSortingOrder(order: SortingOrder) = viewModelScope.launch {
+        userDataRepository.setAppSortingOrder(order)
+    }
+
+    fun updateShowRunningAppsOnTop(shouldShow: Boolean) {
+        viewModelScope.launch {
+            userDataRepository.setShowRunningAppsOnTop(shouldShow)
+        }
     }
 
     private fun appComparator(sortType: AppSorting, sortOrder: SortingOrder): Comparator<AppItem> =
