+++ feature/applist/src/main/java/com/merxury/blocker/feature/applist/navigation/AppListNavigation.kt
@@ -32,7 +32,6 @@ fun NavGraphBuilder.appListScreen(
     navigateToAppDetail: (String) -> Unit,
     navigateToSettings: () -> Unit,
     navigateToSupportAndFeedback: () -> Unit,
-    navigateTooAppSortScreen: () -> Unit,
 ) {
     composable(
         route = appListRoute,
@@ -41,7 +40,6 @@ fun NavGraphBuilder.appListScreen(
             navigateToAppDetail = navigateToAppDetail,
             navigateToSettings = navigateToSettings,
             navigateToSupportAndFeedback = navigateToSupportAndFeedback,
-            navigateTooAppSortScreen = navigateTooAppSortScreen,
         )
     }
 }
