--- feature/sort/src/main/res/values-ru/strings.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-     Copyright 2023 Blocker
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-<resources>
-    <string name="sort_options">Параметры сортировки</string>
-    <string name="sort_by">Сортировать по</string>
-    <string name="package_name">Имя пакета</string>
-    <string name="component_name">Имя компонента</string>
-    <string name="order">порядок</string>
-    <string name="ascending">Восходящий</string>
-    <string name="descending">Нисходящий</string>
-    <string name="priority">Приоритет</string>
-    <string name="disabled_first">Отключенный первый</string>
-    <string name="enabled_first">Включенный первый</string>
-    <string name="none">Нет</string>
-    <string name="name">Имя</string>
-    <string name="install_date">Дата установки</string>
-    <string name="last_updated">Последнее обновление</string>
-    <string name="show_running_apps_on_top">Показывать запущенные приложения сверху</string>
-    <string name="on">включать</string>
-    <string name="off">закрытие</string>
-</resources>
