--- feature/sort/src/main/res/values-zh-rTW/strings.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-     Copyright 2023 Blocker
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-<resources>
-    <string name="sort_options">排序選項</string>
-    <string name="sort_by">排序方式</string>
-    <string name="package_name">包名</string>
-    <string name="component_name">組件名</string>
-    <string name="order">順序</string>
-    <string name="ascending">升序</string>
-    <string name="descending">降序</string>
-    <string name="priority">優先級</string>
-    <string name="disabled_first">禁用組件優先</string>
-    <string name="enabled_first">啟用組件優先</string>
-    <string name="none">無</string>
-    <string name="name">姓名</string>
-    <string name="install_date">安裝日期</string>
-    <string name="last_updated">最後更新</string>
-    <string name="show_running_apps_on_top">置頂運行中應用</string>
-    <string name="on">開啟</string>
-    <string name="off">關閉</string>
-</resources>
