--- feature/sort/src/main/res/values/strings.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-     Copyright 2023 Blocker
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-<resources>
-    <string name="sort_options">Sort options</string>
-    <string name="sort_by">Sort by</string>
-    <string name="package_name">Package name</string>
-    <string name="component_name">Component name</string>
-    <string name="order">Order</string>
-    <string name="ascending">Ascending</string>
-    <string name="descending">Descending</string>
-    <string name="priority">Priority</string>
-    <string name="disabled_first">Disabled first</string>
-    <string name="enabled_first">Enabled first</string>
-    <string name="none">None</string>
-    <string name="name">Name</string>
-    <string name="install_date">Install Date</string>
-    <string name="last_updated">Last Updated</string>
-    <string name="show_running_apps_on_top">Show running apps on top</string>
-    <string name="on">On</string>
-    <string name="off">Off</string>
-</resources>
