+++ settings.gradle.kts
@@ -61,6 +61,5 @@ include(":feature:generalrule")
 include(":feature:ruledetail")
 include(":feature:search")
 include(":feature:settings")
-include(":feature:sort")
 include(":lint")
 include(":ui-test-hilt-manifest")
